Adding authentication
Taiga uses a session based auth where we need to pass in:
 - username
 - password
 - type: 'normal'

This returns an `auth` object that contains an `auth_token`, which
can be used for further authentications.
We use the `beforeRequest` hook to construct the proper `Authorization`
header.

This commit also contains a complete dockerised setup that we can use to
run tests against a `Taiga backend`. This works both locally and in
CircleCI.

To run locally:
```
docker-compose up -d
docker-compose exec taiga-zap sh
zapier test
```

Note: if you're planning to run the cluster without docker, make sure to
export the variables found in `docker-vars.env`.